exit functions migration + various minor fix

+++ src/www/developer/rate.php
@@ -1,23 +1,24 @@
 <?php
 /**
- * GForge rate user page
+ * FusionForge rate user page
  *
  * Copyright 1999-2001 (c) VA Linux Systems
+ * Copyright 2010 (c) Franck Villaume
  *
- * This file is part of GForge.
+ * This file is part of FusionForge.
  *
- * GForge is free software; you can redistribute it and/or modify
+ * FusionForge is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation; either version 2 of the License, or
  * (at your option) any later version.
  *
- * GForge is distributed in the hope that it will be useful,
+ * FusionForge is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with GForge; if not, write to the Free Software
+ * along with FusionForge; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307 USA
  */
 
@@ -26,17 +27,12 @@ require_once $gfcommon.'include/pre.php';
 require_once $gfwww.'include/vote_function.php';
 
 if (!session_loggedin()) {
-
 	exit_not_logged_in();
-
 } else {
 
 	$me =& session_get_user();
 	if (!$me->usesRatings()) {
-		exit_error(
-			_('Ratings turned off'),
-			_('You chose not to participate in the peer rating system')
-		);
+		exit_error(_('Ratings turned off. You chose not to participate in the peer rating system'),'my');
 	}
 
 	$rated_user = getStringFromRequest('rated_user');
@@ -54,40 +50,39 @@ if (!session_loggedin()) {
 			} else {
 				//ratings can only be between +3 and -3
 				if ($rating > 3 || $rating < -3) {
-					$feedback .= _('Invalid rate value');
+					$warning_msg .= _('Invalid rate value');
 				} else {
 					if ($rating) {
 						// get rid of 0.1 thing
 						$rating = (int)$rating;
 						//user did answer this question, so insert into db
-						$res=db_query_params ('SELECT * FROM user_ratings 
-WHERE rated_by=$1 AND user_id=$2 AND rate_field=$3',
-			array($ruser,
-				$rated_user,
-				$i)) ;
+						$res=db_query_params ('SELECT * FROM user_ratings WHERE rated_by=$1 AND user_id=$2 AND rate_field=$3',
+								array($ruser,
+									$rated_user,
+									$i)) ;
 
 						if ($res && db_numrows($res) > 0) {
-							$res=db_query_params ('DELETE FROM user_ratings 
-WHERE rated_by=$1 AND user_id=$2 AND rate_field=$3',
-			array($ruser,
-				$rated_user,
-				$i)) ;
+							$res=db_query_params ('DELETE FROM user_ratings WHERE rated_by=$1 AND user_id=$2 AND rate_field=$3',
+									array($ruser,
+										$rated_user,
+										$i)) ;
 
 						}
-						$res=db_query_params ('INSERT INTO user_ratings (rated_by,user_id,rate_field,rating) 
-VALUES ($1,$2,$3,$4)',
-			array($ruser,
-				$rated_user,
-				$i,
-				$rating)) ;
+						$res=db_query_params ('INSERT INTO user_ratings (rated_by,user_id,rate_field,rating) VALUES ($1,$2,$3,$4)',
+									array($ruser,
+										$rated_user,
+										$i,
+										$rating)) ;
 
-						echo db_error();
+						if(db_error()) {
+							exit_error(db_error(),'my');
+						}
 					}
 				}
 			}
 		}
 	} else {
-		exit_error(_('Error'),_('You can\'t rate yourself'));
+		exit_error(_('You can\'t rate yourself'),'my');
 	}
 
 	echo $HTML->header(array('title'=>_('User Ratings Page')));
@@ -98,7 +93,6 @@ VALUES ($1,$2,$3,$4)',
 	<p>&nbsp;</p>';
 
 	echo $HTML->footer(array());
-
 }
 
 ?>
